feat: Load mandatory data during the splash screen
