+++ lib/src/core/dataloaders/kana_dataloader.dart
@@ -1,0 +1,36 @@
+import 'dart:convert';
+
+import 'package:isar/isar.dart';
+import 'package:kana_to_kanji/src/core/models/kana.dart';
+import 'package:kana_to_kanji/src/core/services/api_service.dart';
+import 'package:kana_to_kanji/src/locator.dart';
+import 'package:http/http.dart' as http;
+
+class KanaDataLoader {
+  final ApiService _apiService = locator<ApiService>();
+  final Isar _isar = locator<Isar>();
+
+  /// Load all the kana from the API.
+  Future loadCollection() async {
+    return _apiService
+        .get('/v1/kanas')
+        .then((response) => _extractKanas(response))
+        .then((listKana) => _isar.write((isar) => isar.kanas.putAll(listKana)));
+  }
+
+  /// Extract all the kana from the API Response.
+  List<Kana> _extractKanas(http.Response response) {
+    if (response.statusCode == 200) {
+      List<Kana> kanas = [];
+      var listKana = jsonDecode(response.body);
+      for (final k in listKana) {
+        kanas.add(Kana.fromJson(k));
+      }
+      return kanas;
+    } else {
+      // If the server did not return a 200 OK response,
+      // then throw an exception.
+      return List.empty();
+    }
+  }
+}
