+++ lib/src/core/services/kana_service.dart
@@ -1,14 +1,9 @@
-import 'dart:convert';
-
 import 'package:isar/isar.dart';
 import 'package:kana_to_kanji/src/core/models/kana.dart';
-import 'package:kana_to_kanji/src/core/services/api_service.dart';
 import 'package:kana_to_kanji/src/locator.dart';
-import 'package:http/http.dart' as http;
 
 class KanaService {
   final Isar _isar = locator<Isar>();
-  final ApiService _apiService = locator<ApiService>();
 
   /// Get all the kana related to the group ids given in parameter.
   Future<List<Kana>> getByGroupIds(List<int> groupIds) async {
@@ -22,17 +17,7 @@ class KanaService {
       kanaQuery = kanaQuery.or().groupIdEqualTo(groupIds[i]);
     }
 
-    final kanas = kanaQuery.findAll();
-
-    if (kanas.isEmpty) {
-      return loadCollection()
-          .then((_) => kanaQuery.findAll())
-          .then((value) async {
-        return value;
-      });
-    }
-
-    return Future.value(kanas);
+    return Future.value(kanaQuery.findAll());
   }
 
   /// Get all the kana related to the group id given in parameter.
@@ -40,27 +25,4 @@ class KanaService {
     return getByGroupIds([groupId]);
   }
 
-  /// Load all the kana from the API.
-  Future<dynamic> loadCollection() {
-    return _apiService
-        .get('/v1/kanas')
-        .then((response) => _extractKanas(response))
-        .then((listKana) => _isar.write((isar) => isar.kanas.putAll(listKana)));
-  }
-
-  /// Extract all the kana from the API Response.
-  List<Kana> _extractKanas(http.Response response) {
-    if (response.statusCode == 200) {
-      List<Kana> kanas = [];
-      var listKana = jsonDecode(response.body);
-      for (final k in listKana) {
-        kanas.add(Kana.fromJson(k));
-      }
-      return kanas;
-    } else {
-      // If the server did not return a 200 OK response,
-      // then throw an exception.
-      return List.empty();
-    }
-  }
 }
